refactor(foundation-demo): restrict access to ShippingContainer to the checkout page
restrict access to ShippingContainer to the checkout page
