+++ apps/foundation-demo/src/app/checkout/pages/checkout-view/checkout-view.component.html
@@ -1,5 +1,11 @@
 <div class="checkout">
   <div class="checkout__shipping" shipping-container #ShippingContainer="ShippingContainer">
-    <shipping-async-wrapper [isShippingInfoValid]="ShippingContainer.isShippingInfoValid$ | async"></shipping-async-wrapper>
+    <shipping 
+      [isShippingInfoValid]="ShippingContainer.isShippingInfoValid$ | async"
+      [shippingInfo]="ShippingContainer.shippingInfo$ | async"
+      [shippingOption]="ShippingContainer.shippingOption$ | async"
+      (updateShippingInfo)="ShippingContainer.updateShippingInfo($event)"
+      (updateShippingOption)="ShippingContainer.updateShippingOption($event)"
+      ></shipping>
   </div>
 </div>
