+++ apps/foundation-demo/src/app/checkout/pages/checkout-view/checkout-view.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   templateUrl: './checkout-view.component.html',
